Move script middleware wiring to pkg/vmcp/cli/serve.go
The upstream refactored cmd/vmcp/app/commands.go to delegate to
pkg/vmcp/cli/serve.go. Move createScriptMiddleware and its wiring
to serve.go to match the current architecture.

Part of #4742
