+++ cmd/vmcp/app/commands.go
@@ -5,46 +5,14 @@
 package app
 
 import (
-	"context"
-	"errors"
 	"fmt"
 	"log/slog"
-	"net/http"
-	"os"
-	"path/filepath"
-	"time"
 
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
-	"go.opentelemetry.io/otel/trace"
-	"gopkg.in/yaml.v3"
-	"k8s.io/client-go/rest"
 
-	"github.com/stacklok/toolhive-core/env"
-	"github.com/stacklok/toolhive/pkg/audit"
-	"github.com/stacklok/toolhive/pkg/auth/upstreamtoken"
-	authserverconfig "github.com/stacklok/toolhive/pkg/authserver"
-	authserverrunner "github.com/stacklok/toolhive/pkg/authserver/runner"
-	"github.com/stacklok/toolhive/pkg/authserver/server/keys"
-	"github.com/stacklok/toolhive/pkg/container/runtime"
-	"github.com/stacklok/toolhive/pkg/groups"
-	"github.com/stacklok/toolhive/pkg/script"
-	"github.com/stacklok/toolhive/pkg/telemetry"
-	"github.com/stacklok/toolhive/pkg/vmcp"
-	"github.com/stacklok/toolhive/pkg/vmcp/aggregator"
-	vmcpauth "github.com/stacklok/toolhive/pkg/vmcp/auth"
-	"github.com/stacklok/toolhive/pkg/vmcp/auth/factory"
-	vmcpclient "github.com/stacklok/toolhive/pkg/vmcp/client"
-	"github.com/stacklok/toolhive/pkg/vmcp/config"
-	"github.com/stacklok/toolhive/pkg/vmcp/discovery"
-	"github.com/stacklok/toolhive/pkg/vmcp/health"
-	"github.com/stacklok/toolhive/pkg/vmcp/k8s"
-	"github.com/stacklok/toolhive/pkg/vmcp/optimizer"
-	vmcprouter "github.com/stacklok/toolhive/pkg/vmcp/router"
-	vmcpserver "github.com/stacklok/toolhive/pkg/vmcp/server"
-	vmcpsession "github.com/stacklok/toolhive/pkg/vmcp/session"
-	"github.com/stacklok/toolhive/pkg/vmcp/session/optimizerdec"
-	vmcpstatus "github.com/stacklok/toolhive/pkg/vmcp/status"
+	"github.com/stacklok/toolhive/pkg/versions"
+	vmcpcli "github.com/stacklok/toolhive/pkg/vmcp/cli"
 )
 
 var rootCmd = &cobra.Command{
@@ -107,7 +75,23 @@ func newServeCmd() *cobra.Command {
 The server will read the configuration file specified by --config flag and start
 listening for MCP client connections. It will aggregate tools, resources, and prompts
 from all configured backend MCP servers.`,
-		RunE: runServe,
+		RunE: func(cmd *cobra.Command, _ []string) error {
+			configPath := viper.GetString("config")
+			if configPath == "" {
+				return fmt.Errorf("no configuration file specified, use --config flag")
+			}
+
+			host, _ := cmd.Flags().GetString("host")
+			port, _ := cmd.Flags().GetInt("port")
+			enableAudit, _ := cmd.Flags().GetBool("enable-audit")
+
+			return vmcpcli.Serve(cmd.Context(), vmcpcli.ServeConfig{
+				ConfigPath:  configPath,
+				Host:        host,
+				Port:        port,
+				EnableAudit: enableAudit,
+			})
+		},
 	}
 
 	// Add serve-specific flags
@@ -125,8 +109,7 @@ func newVersionCmd() *cobra.Command {
 		Short: "Print version information",
 		Long:  "Display version information for vmcp",
 		Run: func(_ *cobra.Command, _ []string) {
-			// Version information will be injected at build time
-			slog.Info(fmt.Sprintf("vmcp version: %s", getVersion()))
+			slog.Info(fmt.Sprintf("vmcp version: %s", versions.Version))
 		},
 	}
 }
@@ -143,534 +126,14 @@ This command checks:
 - Required fields presence
 - Middleware configuration correctness
 - Backend configuration validity`,
-		RunE: func(_ *cobra.Command, _ []string) error {
+		RunE: func(cmd *cobra.Command, _ []string) error {
 			configPath := viper.GetString("config")
 			if configPath == "" {
 				return fmt.Errorf("no configuration file specified, use --config flag")
 			}
-
-			slog.Info(fmt.Sprintf("Validating configuration: %s", configPath))
-
-			// Load configuration from YAML
-			envReader := &env.OSReader{}
-			loader := config.NewYAMLLoader(configPath, envReader)
-			cfg, err := loader.Load()
-			if err != nil {
-				slog.Error(fmt.Sprintf("Failed to load configuration: %v", err))
-				return fmt.Errorf("configuration loading failed: %w", err)
-			}
-
-			slog.Debug("configuration loaded successfully, performing validation")
-
-			// Validate configuration
-			validator := config.NewValidator()
-			if err := validator.Validate(cfg); err != nil {
-				slog.Error(fmt.Sprintf("Configuration validation failed: %v", err))
-				return fmt.Errorf("validation failed: %w", err)
-			}
-
-			slog.Info("✓ Configuration is valid")
-			slog.Info(fmt.Sprintf("  Name: %s", cfg.Name))
-			slog.Info(fmt.Sprintf("  Group: %s", cfg.Group))
-			slog.Info(fmt.Sprintf("  Incoming Auth: %s", cfg.IncomingAuth.Type))
-			slog.Info(fmt.Sprintf("  Outgoing Auth: %s (source: %s)",
-				func() string {
-					if len(cfg.OutgoingAuth.Backends) > 0 {
-						return fmt.Sprintf("%d backends configured", len(cfg.OutgoingAuth.Backends))
-					}
-					return "default only"
-				}(),
-				cfg.OutgoingAuth.Source))
-			slog.Info(fmt.Sprintf("  Conflict Resolution: %s", cfg.Aggregation.ConflictResolution))
-
-			if len(cfg.CompositeTools) > 0 {
-				slog.Info(fmt.Sprintf("  Composite Tools: %d defined", len(cfg.CompositeTools)))
-			}
-
-			return nil
+			return vmcpcli.Validate(cmd.Context(), vmcpcli.ValidateConfig{
+				ConfigPath: configPath,
+			})
 		},
 	}
-}
-
-// getVersion returns the version string (will be set at build time)
-func getVersion() string {
-	// This will be replaced with actual version info using ldflags
-	return "dev"
-}
-
-// getStatusReportingInterval extracts the status reporting interval from config.
-// Returns 0 if not configured, which will use the default interval.
-func getStatusReportingInterval(cfg *config.Config) time.Duration {
-	if cfg.Operational != nil &&
-		cfg.Operational.FailureHandling != nil &&
-		cfg.Operational.FailureHandling.StatusReportingInterval > 0 {
-		return time.Duration(cfg.Operational.FailureHandling.StatusReportingInterval)
-	}
-	return 0
-}
-
-// loadAndValidateConfig loads and validates the vMCP configuration file
-func loadAndValidateConfig(configPath string) (*config.Config, error) {
-	slog.Info(fmt.Sprintf("Loading configuration from: %s", configPath))
-
-	envReader := &env.OSReader{}
-	loader := config.NewYAMLLoader(configPath, envReader)
-	cfg, err := loader.Load()
-	if err != nil {
-		slog.Error(fmt.Sprintf("Failed to load configuration: %v", err))
-		return nil, fmt.Errorf("configuration loading failed: %w", err)
-	}
-
-	validator := config.NewValidator()
-	if err := validator.Validate(cfg); err != nil {
-		slog.Error(fmt.Sprintf("Configuration validation failed: %v", err))
-		return nil, fmt.Errorf("validation failed: %w", err)
-	}
-
-	slog.Info("configuration loaded and validated successfully")
-	slog.Info(fmt.Sprintf("  Name: %s", cfg.Name))
-	slog.Info(fmt.Sprintf("  Group: %s", cfg.Group))
-	slog.Info(fmt.Sprintf("  Conflict Resolution: %s", cfg.Aggregation.ConflictResolution))
-	if len(cfg.CompositeTools) > 0 {
-		slog.Info(fmt.Sprintf("  Composite Tools: %d defined", len(cfg.CompositeTools)))
-	}
-
-	return cfg, nil
-}
-
-// loadAuthServerConfig loads the auth server RunConfig from a sibling file
-// alongside the main config. The operator serializes authserver.RunConfig
-// as a separate ConfigMap key (authserver-config.yaml).
-// Returns nil with no error if the file does not exist.
-func loadAuthServerConfig(configPath string) (*authserverconfig.RunConfig, error) {
-	authServerPath := filepath.Join(filepath.Dir(configPath), "authserver-config.yaml")
-	//nolint:gosec // path derived from trusted --config flag
-	authServerData, readErr := os.ReadFile(authServerPath)
-	if readErr != nil {
-		if errors.Is(readErr, os.ErrNotExist) {
-			return nil, nil
-		}
-		return nil, fmt.Errorf("failed to read auth server config %s: %w", authServerPath, readErr)
-	}
-	var rc authserverconfig.RunConfig
-	if unmarshalErr := yaml.Unmarshal(authServerData, &rc); unmarshalErr != nil {
-		return nil, fmt.Errorf("failed to parse auth server config %s: %w", authServerPath, unmarshalErr)
-	}
-	slog.Info("auth server configuration loaded", "path", authServerPath)
-	return &rc, nil
-}
-
-// discoverBackends initializes managers, discovers backends, and creates backend client
-// Returns empty backends list with no error if running in Kubernetes where CLI discovery doesn't work
-// Also returns the outgoingRegistry for use in creating SessionFactory
-func discoverBackends(
-	ctx context.Context,
-	cfg *config.Config,
-) ([]vmcp.Backend, vmcp.BackendClient, vmcpauth.OutgoingAuthRegistry, error) {
-	// Create outgoing authentication registry
-	slog.Info("initializing outgoing authentication")
-	envReader := &env.OSReader{}
-	outgoingRegistry, err := factory.NewOutgoingAuthRegistry(ctx, envReader)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("failed to create outgoing authentication registry: %w", err)
-	}
-
-	// Create backend client first (needed even with empty backends)
-	backendClient, err := vmcpclient.NewHTTPBackendClient(outgoingRegistry)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("failed to create backend client: %w", err)
-	}
-
-	// Create backend discoverer based on configuration mode
-	var discoverer aggregator.BackendDiscoverer
-	if len(cfg.Backends) > 0 {
-		// Static mode: Use pre-configured backends from config (no K8s API access needed)
-		slog.Info(fmt.Sprintf("Static mode: using %d pre-configured backends", len(cfg.Backends)))
-		discoverer = aggregator.NewUnifiedBackendDiscovererWithStaticBackends(
-			cfg.Backends,
-			cfg.OutgoingAuth,
-			cfg.Group,
-		)
-	} else {
-		// Dynamic mode: Discover backends at runtime from K8s API
-		slog.Info("dynamic mode: initializing group manager for backend discovery")
-		groupsManager, err := groups.NewManager()
-		if err != nil {
-			return nil, nil, nil, fmt.Errorf("failed to create groups manager: %w", err)
-		}
-
-		discoverer, err = aggregator.NewBackendDiscoverer(ctx, groupsManager, cfg.OutgoingAuth)
-		if err != nil {
-			return nil, nil, nil, fmt.Errorf("failed to create backend discoverer: %w", err)
-		}
-	}
-
-	slog.Info(fmt.Sprintf("Discovering backends in group: %s", cfg.Group))
-	backends, err := discoverer.Discover(ctx, cfg.Group)
-	if err != nil {
-		return nil, nil, nil, fmt.Errorf("failed to discover backends: %w", err)
-	}
-
-	if len(backends) == 0 {
-		slog.Warn(fmt.Sprintf("No backends discovered in group %s - vmcp will start but have no backends to proxy", cfg.Group))
-		return []vmcp.Backend{}, backendClient, outgoingRegistry, nil
-	}
-
-	slog.Info(fmt.Sprintf("Discovered %d backends", len(backends)))
-	return backends, backendClient, outgoingRegistry, nil
-}
-
-// createSessionFactory creates a MultiSessionFactory with HMAC-SHA256 token binding.
-// The HMAC secret is read from the VMCP_SESSION_HMAC_SECRET environment variable.
-//
-// Behavior:
-//   - If VMCP_SESSION_HMAC_SECRET is set: validates length and creates factory with the secret
-//   - If running in Kubernetes without secret: returns error (production safety requirement)
-//   - Otherwise: logs warning and creates factory with default insecure secret
-//
-// Returns an error only when running in Kubernetes without a secret (fail-fast for production).
-func createSessionFactory(
-	outgoingRegistry vmcpauth.OutgoingAuthRegistry,
-	agg aggregator.Aggregator,
-) (vmcpsession.MultiSessionFactory, error) {
-	const (
-		envKey                  = "VMCP_SESSION_HMAC_SECRET"
-		minRecommendedSecretLen = 32
-	)
-
-	// Build base options, conditionally including aggregator only when provided.
-	opts := []vmcpsession.MultiSessionFactoryOption{}
-	if agg != nil {
-		opts = append(opts, vmcpsession.WithAggregator(agg))
-	}
-
-	hmacSecret := os.Getenv(envKey)
-
-	if hmacSecret != "" {
-		// Validate secret length (warn if too short, but still use it)
-		if secretLen := len(hmacSecret); secretLen < minRecommendedSecretLen {
-			// G706: Safe - only logging integer length, not the secret itself
-			slog.Warn( //nolint:gosec
-				"HMAC secret is shorter than recommended length - consider using a longer secret",
-				"actual_length", secretLen,
-				"recommended_length", minRecommendedSecretLen,
-			)
-		}
-		slog.Info("using HMAC secret from VMCP_SESSION_HMAC_SECRET environment variable for session token binding")
-		opts = append(opts, vmcpsession.WithHMACSecret([]byte(hmacSecret)))
-		return vmcpsession.NewSessionFactory(outgoingRegistry, opts...), nil
-	}
-
-	// No secret provided - check if we're in Kubernetes (production environment)
-	if runtime.IsKubernetesRuntime() {
-		return nil, fmt.Errorf(
-			"VMCP_SESSION_HMAC_SECRET environment variable is required when running in Kubernetes. " +
-				"Generate a secure secret with: openssl rand -base64 32",
-		)
-	}
-
-	// Development mode: use default insecure secret with warning
-	slog.Warn("VMCP_SESSION_HMAC_SECRET not set - using default insecure secret (NOT recommended for production)")
-	return vmcpsession.NewSessionFactory(outgoingRegistry, opts...), nil
-}
-
-// runServe implements the serve command logic
-//
-//nolint:gocyclo // Complexity from server initialization and configuration is acceptable
-func runServe(cmd *cobra.Command, _ []string) error {
-	ctx := cmd.Context()
-	configPath := viper.GetString("config")
-
-	if configPath == "" {
-		return fmt.Errorf("no configuration file specified, use --config flag")
-	}
-
-	// Load and validate configuration
-	cfg, err := loadAndValidateConfig(configPath)
-	if err != nil {
-		return err
-	}
-
-	// Check if --enable-audit flag is set
-	enableAudit, _ := cmd.Flags().GetBool("enable-audit")
-	if enableAudit && cfg.Audit == nil {
-		// Create default audit config with reasonable defaults
-		cfg.Audit = audit.DefaultConfig()
-		cfg.Audit.Component = "vmcp-server"
-		slog.Info("audit logging enabled with default configuration")
-	}
-
-	// Load auth server config from sibling file if present.
-	authServerRC, err := loadAuthServerConfig(configPath)
-	if err != nil {
-		return err
-	}
-
-	// Auto-populate SubjectProviderName on any token_exchange strategy that
-	// omitted it when an embedded auth server is active.  Without this, the
-	// strategy silently falls back to identity.Token (the ToolHive JWT) as the
-	// RFC 8693 subject token, which the exchange endpoint rejects.
-	config.InjectSubjectProviderNames(cfg, authServerRC)
-
-	// Construct embedded authorization server if configured.
-	// Hard failure — if the auth server is configured but cannot be
-	// created, we must not silently fall back to unauthenticated mode.
-	var embeddedAuthServer *authserverrunner.EmbeddedAuthServer
-	if authServerRC != nil {
-		embeddedAuthServer, err = authserverrunner.NewEmbeddedAuthServer(ctx, authServerRC)
-		if err != nil {
-			return fmt.Errorf("failed to create embedded auth server: %w", err)
-		}
-		defer func() {
-			if closeErr := embeddedAuthServer.Close(); closeErr != nil {
-				slog.Error(fmt.Sprintf("failed to close embedded auth server: %v", closeErr))
-			}
-		}()
-		slog.Info("embedded authorization server initialized")
-	}
-
-	// Discover backends and create client
-	backends, backendClient, outgoingRegistry, err := discoverBackends(ctx, cfg)
-	if err != nil {
-		return err
-	}
-
-	// Create conflict resolver based on configuration
-	conflictResolver, err := aggregator.NewConflictResolver(cfg.Aggregation)
-	if err != nil {
-		return fmt.Errorf("failed to create conflict resolver: %w", err)
-	}
-
-	// If telemetry is configured, create the provider early so aggregator can use it
-	var telemetryProvider *telemetry.Provider
-	if cfg.Telemetry != nil {
-		telemetryProvider, err = telemetry.NewProvider(ctx, *cfg.Telemetry)
-		if err != nil {
-			return fmt.Errorf("failed to create telemetry provider: %w", err)
-		}
-		defer func() {
-			err := telemetryProvider.Shutdown(ctx)
-			if err != nil {
-				slog.Error(fmt.Sprintf("failed to shutdown telemetry provider: %v", err))
-			}
-		}()
-	}
-
-	// Create aggregator with tracer provider (nil if telemetry not configured)
-	var tracerProvider trace.TracerProvider
-	if telemetryProvider != nil {
-		tracerProvider = telemetryProvider.TracerProvider()
-	}
-	agg := aggregator.NewDefaultAggregator(backendClient, conflictResolver, cfg.Aggregation, tracerProvider)
-
-	// DynamicRegistry tracks backends for dynamic discovery in Kubernetes mode.
-	// In standalone mode: backends from config file, no dynamic updates.
-	// In K8s mode with discovered auth: backends watched dynamically via BackendWatcher.
-	dynamicRegistry := vmcp.NewDynamicRegistry(backends)
-	backendRegistry := vmcp.BackendRegistry(dynamicRegistry)
-
-	discoveryMgr, err := discovery.NewManager(agg)
-	if err != nil {
-		return fmt.Errorf("failed to create discovery manager: %w", err)
-	}
-	slog.Info("dynamic backend registry enabled for Kubernetes environment")
-
-	// Backend watcher for dynamic backend discovery
-	var backendWatcher *k8s.BackendWatcher
-
-	// If outgoingAuth.source is "discovered", start K8s backend watcher to watch backend changes
-	if cfg.OutgoingAuth != nil && cfg.OutgoingAuth.Source == "discovered" {
-		slog.Info("detected dynamic backend discovery mode (outgoingAuth.source: discovered)")
-
-		// Get in-cluster REST config
-		restConfig, err := rest.InClusterConfig()
-		if err != nil {
-			return fmt.Errorf("failed to get in-cluster config: %w", err)
-		}
-
-		// Get namespace from environment variable set by operator
-		// The operator sets VMCP_NAMESPACE to the VirtualMCPServer's namespace
-		namespace := os.Getenv("VMCP_NAMESPACE")
-		if namespace == "" {
-			return fmt.Errorf("VMCP_NAMESPACE environment variable not set")
-		}
-
-		// Create K8s backend watcher to watch backend changes
-		backendWatcher, err = k8s.NewBackendWatcher(restConfig, namespace, cfg.Group, dynamicRegistry)
-		if err != nil {
-			return fmt.Errorf("failed to create backend watcher: %w", err)
-		}
-
-		// Start K8s backend watcher in background goroutine
-		go func() {
-			slog.Info("starting Kubernetes backend watcher in background")
-			if err := backendWatcher.Start(ctx); err != nil {
-				slog.Error(fmt.Sprintf("Backend watcher stopped with error: %v", err))
-			}
-		}()
-
-		slog.Info("kubernetes backend watcher started for dynamic backend discovery")
-	}
-
-	// Create router
-	rtr := vmcprouter.NewDefaultRouter()
-
-	slog.Info(fmt.Sprintf("Setting up incoming authentication (type: %s)", cfg.IncomingAuth.Type))
-
-	// Create server configuration with flags
-	// Cobra validates flag types at parse time, so these values are safe to use directly
-	host, _ := cmd.Flags().GetString("host")
-	port, _ := cmd.Flags().GetInt("port")
-
-	// Note: telemetryProvider was already created earlier (before aggregator creation)
-	// to enable tracing in the aggregator
-
-	// Configure health monitoring if enabled
-	var healthMonitorConfig *health.MonitorConfig
-	if cfg.Operational != nil && cfg.Operational.FailureHandling != nil && cfg.Operational.FailureHandling.HealthCheckInterval > 0 {
-		// Note: HealthCheckInterval is config.Duration (alias for time.Duration), already in nanoseconds
-		// from YAML/JSON parsing via time.ParseDuration. This is a simple type cast, not unit conversion.
-		checkInterval := time.Duration(cfg.Operational.FailureHandling.HealthCheckInterval)
-		if cfg.Operational.FailureHandling.UnhealthyThreshold < 1 {
-			return fmt.Errorf("invalid health check configuration: unhealthy threshold must be >= 1, got %d",
-				cfg.Operational.FailureHandling.UnhealthyThreshold)
-		}
-
-		defaults := health.DefaultConfig()
-
-		// Use configured timeout if provided, otherwise use default
-		healthCheckTimeout := defaults.Timeout
-		if cfg.Operational.FailureHandling.HealthCheckTimeout > 0 {
-			healthCheckTimeout = time.Duration(cfg.Operational.FailureHandling.HealthCheckTimeout)
-		}
-
-		healthMonitorConfig = &health.MonitorConfig{
-			CheckInterval:      checkInterval,
-			UnhealthyThreshold: cfg.Operational.FailureHandling.UnhealthyThreshold,
-			Timeout:            healthCheckTimeout,
-			DegradedThreshold:  defaults.DegradedThreshold,
-		}
-
-		// Wire circuit breaker configuration if present
-		if cfg.Operational.FailureHandling.CircuitBreaker != nil {
-			cbConfig := cfg.Operational.FailureHandling.CircuitBreaker
-
-			healthMonitorConfig.CircuitBreaker = &health.CircuitBreakerConfig{
-				Enabled:          cbConfig.Enabled,
-				FailureThreshold: cbConfig.FailureThreshold,
-				Timeout:          time.Duration(cbConfig.Timeout),
-			}
-
-			if cbConfig.Enabled {
-				slog.Info(fmt.Sprintf("Circuit breaker enabled (threshold: %d failures, timeout: %v)",
-					cbConfig.FailureThreshold, time.Duration(cbConfig.Timeout)))
-			}
-		}
-
-		slog.Info("health monitoring configured from operational settings")
-	}
-
-	// Create status reporter using factory (auto-detects K8s vs CLI mode)
-	statusReporter, err := vmcpstatus.NewReporter()
-	if err != nil {
-		return fmt.Errorf("failed to create status reporter: %w", err)
-	}
-
-	optCfg, err := optimizer.GetAndValidateConfig(cfg.Optimizer)
-	if err != nil {
-		return fmt.Errorf("failed to validate optimizer config: %w", err)
-	}
-
-	sessionFactory, err := createSessionFactory(outgoingRegistry, agg)
-	if err != nil {
-		return err
-	}
-
-	// When the optimizer is enabled, its meta-tools (find_tool, call_tool) must pass
-	// through the authz response filter so they appear in tools/list. Authorization
-	// for the underlying backend tools is enforced by the middleware's call_tool
-	// interception. See: https://github.com/stacklok/toolhive/issues/4373
-	var passThroughTools map[string]struct{}
-	if optCfg != nil {
-		passThroughTools = map[string]struct{}{
-			optimizerdec.FindToolName: {},
-			optimizerdec.CallToolName: {},
-		}
-	}
-
-	// Extract dependencies from the embedded auth server so the OIDC middleware
-	// can (a) resolve JWKS keys in-process instead of self-referential HTTP
-	// calls, and (b) enrich Identity with upstream provider tokens.
-	var upstreamReader upstreamtoken.TokenReader
-	var keyProvider keys.PublicKeyProvider
-	if embeddedAuthServer != nil {
-		stor := embeddedAuthServer.IDPTokenStorage()
-		refresher := embeddedAuthServer.UpstreamTokenRefresher()
-		upstreamReader = upstreamtoken.NewInProcessService(stor, refresher)
-		keyProvider = embeddedAuthServer.KeyProvider()
-	}
-
-	authMiddleware, authzMiddleware, authInfoHandler, err :=
-		factory.NewIncomingAuthMiddleware(ctx, cfg.IncomingAuth, passThroughTools, upstreamReader, keyProvider)
-	if err != nil {
-		return fmt.Errorf("failed to create authentication middleware: %w", err)
-	}
-
-	slog.Info(fmt.Sprintf("Incoming authentication configured: %s", cfg.IncomingAuth.Type))
-
-	serverCfg := &vmcpserver.Config{
-		Name:                    cfg.Name,
-		Version:                 getVersion(),
-		GroupRef:                cfg.Group,
-		Host:                    host,
-		Port:                    port,
-		AuthMiddleware:          authMiddleware,
-		AuthzMiddleware:         authzMiddleware,
-		AuthInfoHandler:         authInfoHandler,
-		AuthServer:              embeddedAuthServer,
-		TelemetryProvider:       telemetryProvider,
-		AuditConfig:             cfg.Audit,
-		HealthMonitorConfig:     healthMonitorConfig,
-		StatusReportingInterval: getStatusReportingInterval(cfg),
-		Watcher:                 backendWatcher,
-		StatusReporter:          statusReporter,
-		OptimizerConfig:         optCfg,
-		SessionFactory:          sessionFactory,
-		SessionStorage:          cfg.SessionStorage,
-		ScriptMiddleware:        createScriptMiddleware(cfg),
-	}
-
-	// Convert composite tool configurations to workflow definitions
-	workflowDefs, err := vmcpserver.ConvertConfigToWorkflowDefinitions(cfg.CompositeTools)
-	if err != nil {
-		return fmt.Errorf("failed to convert composite tool definitions: %w", err)
-	}
-	if len(workflowDefs) > 0 {
-		slog.Info(fmt.Sprintf("Loaded %d composite tool workflow definitions", len(workflowDefs)))
-	}
-
-	// Create server with discovery manager, backend registry, and workflow definitions
-	srv, err := vmcpserver.New(ctx, serverCfg, rtr, backendClient, discoveryMgr, backendRegistry, workflowDefs)
-	if err != nil {
-		return fmt.Errorf("failed to create Virtual MCP Server: %w", err)
-	}
-
-	// Start server (blocks until shutdown signal)
-	slog.Info(fmt.Sprintf("Starting Virtual MCP Server at %s", srv.Address()))
-	return srv.Start(ctx)
-}
-
-// createScriptMiddleware returns the script middleware function if code mode
-// is enabled in config, or nil if disabled.
-func createScriptMiddleware(cfg *config.Config) func(http.Handler) http.Handler {
-	if cfg.ScriptEngine == nil || !cfg.ScriptEngine.Enabled {
-		return nil
-	}
-
-	slog.Info("code mode enabled — execute_tool_script virtual tool will be available")
-	return script.NewMiddleware(&script.Config{
-		StepLimit:     cfg.ScriptEngine.StepLimit,
-		ParallelMax:   cfg.ScriptEngine.ParallelMax,
-		ScriptTimeout: time.Duration(cfg.ScriptEngine.ScriptTimeout),
-	})
 }
